perf: ip3561q current_ma 字段 int32_t 缩为 int16_t
int16_t 范围 ±32768mA，移动电源电流监测足够，节省 2 字节结构体空间。

+++ drv/ip3561q.h
@@ -111,7 +111,7 @@ typedef struct
     uint16_t vbat_mv;       /* 电池包总电压 mV (VBAT ADC) */
     uint32_t rntc1_ohm;      /* NTC1 阻值 Ω (IP3561Q NTC1) */
     uint32_t rntc2_ohm;     /* NTC2 阻值 Ω (IP3561Q NTC2) */
-    int32_t  current_ma;    /* 电流 mA (正=充电) */
+    int16_t  current_ma;    /* 电流 mA (正=充电) */
 } ip3561q_proc_data_t;
 
 #define IP3561Q_CELL_CNT    4U   /* 4 串电芯 */
